requirements.txt
fastapi>=0.73.0
starlette>=0.18.0
uvicorn==0.17.1
python-multipart==0.0.5

pydantic>=1.9.0

tortoise-orm==0.18.1
rom>=1.0.4

passlib>=1.7.4
python-jose>=3.3.0
